Update README.md
The gradient operator in Image Processing and Computer Vision (IPCV) is a tool used to detect changes in the intensity or color of an image. It highlights areas where there are sharp transitions, such as edges, by calculating the rate of change of pixel values in both horizontal and vertical directions. When applied, it creates a new image that shows the direction and strength of these changes, making it useful for tasks like edge detection. Essentially, the gradient operator helps in identifying important features, like boundaries, in an image that separate different regions.